requirements.txt
flask==3.0.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
lyricsgenius==3.0.1
flask-cors==4.0.0
httpx==0.27.2
urllib3<2.0
spotipy==2.25.2
beautifulsoup4==4.14.2
